requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
click==8.2.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pydantic==2.11.4
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
PyMuPDF==1.23.14
python-pptx==0.6.21
requests==2.31.0
python-dotenv==1.0.0
